Remove kernel module dependency from systemd
The systemd recipe of poky has some kernel module dependency.  This dependency cause
some kernel module and kernel image installing.

In host case, that dependency is required.
On the other hand, that dependency is unnecessarily in guest case.
Because the kernel space software is required in build time by some board specific
packages.  On the other hand, the kernel and kernel module binary is not required
in runtime, because these package is already installed in host.

This patch remove some kernel module dependency from systemd in guest building.

Signed-off-by: Naoto Yamaguchi <naoto.yamaguchi@aisin.co.jp>