Importing the navigation demo app
The mominavi is a navigation example based on mapbox.  It's based on
aglqtnavigation, that initial version of existing AGL demo navigation app.

Existing agl-appfw is difficult to integrate into linux container guest,
because agl-appfw strongly depend SMACK label in host.
The mominavi is not require agl-appfw.
It's temporally solution for MM release.

This patch import mominavi to create ivi guest for demo.

Bug-AGL: SPEC-4095

Signed-off-by: Naoto Yamaguchi <naoto.yamaguchi@aisin.co.jp>